pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "pipeline_check"
version = "0.0.1"
authors = [
  { name="Andrei Egorov", email="anegorov777@gmail.com" },
]
description = "A small example package"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-gitlab",
    "prometheus_client"
]

[project.scripts]
pipeline-check = "pipeline_check.main:main"

[tool.setuptools.packages.find]
# All the following settings are optional:
where = ["src"]